fix: reduced number of files per command module
