+++ cmd/fullstack-network-manager/src/commands/init/index.mjs
@@ -1,3 +1,11 @@
-import { InitCmd } from './init_cmd.mjs'
+import { Init } from './init.mjs'
+const InitCmd= {
+    command: "init",
+    desc: "Initialize local environment",
+    builder: {},
+    handler: function (argv) {
+        Init.init(argv)
+    },
+}
 
 export { InitCmd }
