--- cmd/fullstack-network-manager/src/commands/init/init_cmd.mjs
@@ -1,9 +1,0 @@
-import { Init } from './init.mjs'
-export const InitCmd= {
-    command: "init",
-    desc: "Initialize local environment",
-    builder: {},
-    handler: function (argv) {
-        Init.init(argv)
-    },
-}
